refactor: replace custom Input component with Shadcn UI Input
- Removed the custom Input component and updated imports across multiple files to utilize the Shadcn UI Input component for consistency.
- Enhanced the overall structure of the components by aligning with the new UI framework, improving maintainability and accessibility.
