+++ frontend/components/EventMetaEditor.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from "react";
 import { Card } from "./ui/card";
-import Input from "./Input";
+import { Input } from "@/components/ui/input";
 import { Button } from "@/components/ui/button";
 import { Sheet, SheetTrigger, SheetContent, SheetHeader, SheetTitle } from "@/components/ui/sheet";
 import { PencilIcon, CheckIcon, XMarkIcon, LinkIcon, ArrowTopRightOnSquareIcon } from "@heroicons/react/24/outline";
