+++ frontend/pages/login.tsx
@@ -4,7 +4,7 @@ import { useRouter } from 'next/router';
 import { UserContext } from './_app';
 import { Button } from "@/components/ui/button";
 import { Card } from "../components/ui/card";
-import Input from '../components/Input';
+import { Input } from "@/components/ui/input";
 
 // Define User interface
 interface User {
